Fixed bugs and added dto

+++ studentManagement/src/main/java/com/uma/southdevelopers/controllers/InstituteController.java
@@ -3,8 +3,11 @@ package com.uma.southdevelopers.controllers;
 
 import com.uma.southdevelopers.dtos.InstituteDTO;
 import com.uma.southdevelopers.entities.Institute;
+import com.uma.southdevelopers.entities.Student;
 import com.uma.southdevelopers.service.InstituteDBService;
+import com.uma.southdevelopers.service.StudentDBService;
 import com.uma.southdevelopers.service.exceptions.EntityNotFoundException;
+import com.uma.southdevelopers.service.exceptions.InstitutoDoNotDeleteException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,9 +23,11 @@ import java.util.function.Function;
 public class InstituteController {
 
     public InstituteDBService service;
+    public StudentDBService studentService;
 
-    public InstituteController(InstituteDBService service) {
+    public InstituteController(InstituteDBService service, StudentDBService studentService) {
         this.service = service;
+        this.studentService = studentService;
     }
 
     public static Function<Long, URI> instituteUriBuilder(UriComponents uriBuilder) {
@@ -44,12 +49,11 @@ public class InstituteController {
     }
 
     @PostMapping
-    public ResponseEntity<?> addInstitute(@RequestBody InstituteDTO institute, UriComponentsBuilder uriBuilder) {
+    public Institute addInstitute(@RequestBody InstituteDTO institute, UriComponentsBuilder uriBuilder) {
         Institute inst = institute.institute();
         Long id = service.addInstitute(inst);
-        return ResponseEntity.created(
-                        instituteUriBuilder(uriBuilder.build()).apply(id))
-                .build();
+        inst.setId(id);
+        return inst;
     }
 
     @GetMapping("/{id}")
@@ -62,19 +66,32 @@ public class InstituteController {
 
     @PutMapping("/{id}")
     @ResponseStatus(code=HttpStatus.OK)
-    public void updateInstitute(@PathVariable Long id, @RequestBody InstituteDTO institute) {
+    public Institute updateInstitute(@PathVariable Long id, @RequestBody InstituteDTO institute) {
         Institute entidadInstituto = institute.institute();
         entidadInstituto.setId(id);
         service.updateInstitute(entidadInstituto);
+        return entidadInstituto;
     }
 
     @DeleteMapping("/{id}")
     public void deleteInstitute(@PathVariable Long id) {
+        Institute institute = service.obtainInstitute(id);
+
+        List<Student> students = studentService.obtainStudentFromInstitute(institute);
+
+        if(students.size() > 0) {
+            throw new InstitutoDoNotDeleteException();
+        }
+
         service.deleteInstitute(id);
     }
 
     @ExceptionHandler(EntityNotFoundException.class)
     @ResponseStatus(code = HttpStatus.NOT_FOUND)
     public void notFound() {}
+
+    @ExceptionHandler(InstitutoDoNotDeleteException.class)
+    @ResponseStatus(code = HttpStatus.CONFLICT)
+    public void instituteWithStudents() {}
 
 }
